requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
chromadb
python-dotenv
pydantic
python-multipart
pypdf
openai
python-dotenv
